Move barcode code to one package
