+++ collect_app/src/main/java/org/odk/collect/android/widgets/WidgetAnswer.kt
@@ -5,6 +5,7 @@ import androidx.compose.ui.Modifier
 import androidx.lifecycle.ViewModelProvider
 import org.javarosa.core.model.Constants
 import org.javarosa.form.api.FormEntryPrompt
+import org.odk.collect.android.widgets.barcode.BarcodeWidgetAnswer
 import org.odk.collect.android.widgets.video.VideoWidgetAnswer
 
 @Composable
@@ -20,7 +21,12 @@ fun WidgetAnswer(
         when (prompt.controlType) {
             Constants.CONTROL_INPUT -> {
                 when (prompt.dataType) {
-                    Constants.DATATYPE_BARCODE -> BarcodeWidgetAnswer(modifier, answer, fontSize, onLongClick)
+                    Constants.DATATYPE_BARCODE -> BarcodeWidgetAnswer(
+                        modifier,
+                        answer,
+                        fontSize,
+                        onLongClick
+                    )
                 }
             }
             Constants.CONTROL_VIDEO_CAPTURE -> VideoWidgetAnswer(modifier, answer, viewModelProvider!!, onLongClick)
