+++ collect_app/src/main/java/org/odk/collect/android/widgets/barcode/BarcodeWidgetAnswer.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.widgets
+package org.odk.collect.android.widgets.barcode
 
 import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.layout.Arrangement
