+++ collect_app/src/main/java/org/odk/collect/android/widgets/barcode/BarcodeWidgetContent.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.widgets
+package org.odk.collect.android.widgets.barcode
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.MaterialTheme
@@ -8,6 +8,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import org.odk.collect.icons.R
+import org.odk.collect.android.widgets.WidgetIconButton
 import org.odk.collect.strings.R.string
 
 @Composable
